+++ tencentcloud/extension_ssl.go
@@ -1,0 +1,8 @@
+package tencentcloud
+
+const (
+	SSL_MODULE_TYPE = "ssl"
+	SSL_WITH_CERT   = "1"
+
+	SSL_STATUS_AVAILABLE = 1
+)